fix bugs: fuzzy search of project

+++ src/main/java/org/blank/projectmanagementsystem/service/impl/TaskServiceImpl.java
@@ -2,7 +2,9 @@ package org.blank.projectmanagementsystem.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.blank.projectmanagementsystem.domain.Enum.ProjectStatus;
 import org.blank.projectmanagementsystem.domain.entity.Phase;
+import org.blank.projectmanagementsystem.domain.entity.Project;
 import org.blank.projectmanagementsystem.domain.entity.Task;
 import org.blank.projectmanagementsystem.domain.entity.User;
 import org.blank.projectmanagementsystem.domain.formInput.TaskFormInput;
@@ -13,6 +15,7 @@ import org.blank.projectmanagementsystem.repository.TaskRepository;
 import org.blank.projectmanagementsystem.repository.UserRepository;
 import org.blank.projectmanagementsystem.service.TaskService;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,9 +34,9 @@ public class TaskServiceImpl implements TaskService {
 
     private final TaskMapper taskMapper = new TaskMapper();
 
-    private User getCurrentUser(){
+    private User getCurrentUser() {
         var username = SecurityContextHolder.getContext().getAuthentication().getName();
-        return userRepository.findByUsernameOrEmail(username,username).orElse(null);
+        return userRepository.findByUsernameOrEmail(username, username).orElse(null);
     }
 
     @Override
@@ -44,14 +47,19 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
+    @Transactional
     public TaskViewObject createTask(TaskFormInput taskFormInput) {
         Task task = taskMapper.mapToTask(taskFormInput);
-        // Save the task to the database
-        Task savedTask = taskRepository.save(fillTaskData(taskFormInput, task));
-        // Map and return the saved task as a TaskViewObject
+        Task resultTask = fillTaskData(taskFormInput, task);
+
+        if(!isAuthorized(resultTask.getProject())){
+            return null;
+        }
+
+        Task savedTask = taskRepository.save(resultTask);
+        checkProjectStatus(savedTask.getProject());
         return taskMapper.mapToTaskViewObject(savedTask);
     }
-
 
     @Override
     @Transactional
@@ -60,50 +68,58 @@ public class TaskServiceImpl implements TaskService {
         //Set phase and project if it exists
         task.setId(taskFormInput.getId());
         var modifyTask = fillTaskData(taskFormInput, task);
+
+        if(!isAuthorized(modifyTask.getProject())){
+            return null;
+        }
         //reset subtask date
         resetSubTaskDate(modifyTask);
         var resultTask = taskRepository.save(modifyTask);
-        if(modifyTask.isStatus()){
+        if (modifyTask.isStatus()) {
             makeChildComplete(modifyTask);
             //check if siblings of the task are completed, then set the parent task to complete
-            if (modifyTask.getParentTask() != null) {
-                var siblings = taskRepository.findAllByParentTask(modifyTask.getParentTask());
-                if (siblings.stream().allMatch(Task::isStatus)) {
+            var siblings = taskRepository.findAllByParentTask(modifyTask.getParentTask());
+
+            if (siblings.stream().allMatch(Task::isStatus)) {
+                if (modifyTask.getParentTask() != null) {
                     modifyTask.getParentTask().setStatus(true);
                     modifyTask.getParentTask().setActualHours(siblings.stream().reduce(0f, (acc, val) -> acc + val.getActualHours(), Float::sum));
                     modifyTask.getParentTask().setActualDueDate(modifyTask.getActualDueDate());
                 }
             }
-        }else {
+
+        } else {
             makeChildIncomplete(modifyTask);
             //check if this task have parent and siblings, then set the parent task to incomplete
-            if (modifyTask.getParentTask() != null) {
-                var siblings = taskRepository.findAllByParentTask(modifyTask.getParentTask());
-                if (siblings.stream().anyMatch(t->!t.isStatus())) {
+
+            var siblings = taskRepository.findAllByParentTask(modifyTask.getParentTask());
+            if (siblings.stream().anyMatch(t -> !t.isStatus())) {
+                if (modifyTask.getParentTask() != null) {
                     modifyTask.getParentTask().setStatus(false);
                     modifyTask.getParentTask().setActualHours(0f);
                     modifyTask.getParentTask().setActualDueDate(null);
                 }
             }
         }
+        checkProjectStatus(resultTask.getProject());
         return resultTask;
     }
 
-    private void makeChildComplete(Task task){
+    private void makeChildComplete(Task task) {
         AtomicReference<Float> parentTaskActualHours = new AtomicReference<>(task.getActualHours());
         var subTasks = taskRepository.findAllByParentTask(task);
         subTasks.stream().filter(Task::isStatus).forEach(val -> {
             parentTaskActualHours.updateAndGet(v -> v - val.getActualHours());
         });
-        subTasks.stream().filter(t->!t.isStatus()).forEach(val -> {
+        subTasks.stream().filter(t -> !t.isStatus()).forEach(val -> {
             val.setStatus(true);
-            val.setActualHours(parentTaskActualHours.get()/subTasks.size());
+            val.setActualHours(parentTaskActualHours.get() / subTasks.size());
             val.setActualDueDate(task.getActualDueDate());
             makeChildComplete(val);
         });
     }
 
-    private void makeChildIncomplete(Task task){
+    private void makeChildIncomplete(Task task) {
         var subTasks = taskRepository.findAllByParentTask(task);
         subTasks.forEach(val -> {
             val.setStatus(false);
@@ -134,9 +150,6 @@ public class TaskServiceImpl implements TaskService {
         return task;
     }
 
-    // recursive function that
-    // reset start date and end date of subtasks as parent task's start date and end date
-    // if subtask start date or end date is not in range of parent task
     private void resetSubTaskDate(Task task) {
         var subTasks = taskRepository.findAllByParentTask(task);
         subTasks.forEach(val -> {
@@ -157,11 +170,13 @@ public class TaskServiceImpl implements TaskService {
         //check task exist or not
         var currentTask = taskRepository.findById(id);
         currentTask.ifPresent(task -> {
-            task.getAssignees().clear();
-            clearAssignees(task);
+            if(isAuthorized(task.getProject())) {
+                task.getAssignees().clear();
+                clearAssignees(task);
+                taskRepository.deleteById(id);
+                checkProjectStatus(task.getProject());
+            }
         });
-
-        taskRepository.deleteById(id);
     }
 
     @Override
@@ -176,8 +191,12 @@ public class TaskServiceImpl implements TaskService {
                 .stream().map(taskMapper::mapToTaskViewObject).toList();
     }
 
+    @Override
+    public Task getTaskById(Long id) {
+        return taskRepository.getReferenceById(id);
+    }
 
-    //create recursive function to get all subtask and clear assignees
+
     private void clearAssignees(Task task) {
         task.getAssignees().clear();
         var subTasks = taskRepository.findAllByParentTask(task);
@@ -185,6 +204,23 @@ public class TaskServiceImpl implements TaskService {
             val.getAssignees().clear();
             clearAssignees(val);
         });
+    }
+
+    private void checkProjectStatus(Project project){
+        var tasks = taskRepository.findAllByProjectId(project.getId());
+        if(project.getStatus()!=ProjectStatus.PENDING){
+            if(tasks.size()==0){
+                project.setStatus(ProjectStatus.ONGOING);
+            }else if(tasks.stream().allMatch(Task::isStatus)){
+                project.setStatus(ProjectStatus.FINISHED);
+            }else {
+                project.setStatus(ProjectStatus.ONGOING);
+            }
+        }
+    }
+
+    private boolean isAuthorized(Project project){
+        return project.getStatus()!=ProjectStatus.PENDING;
     }
 }
